sync: rename `sumneko_lua` to `lua_ls`
