+++ lua/modules/configs/completion/lsp.lua
@@ -37,8 +37,9 @@ return function()
 			"bashls",
 			"clangd",
 			"gopls",
+			"html",
+			"lua_ls",
 			"pyright",
-			"sumneko_lua",
 		},
 	})
 
@@ -96,10 +97,10 @@ return function()
 			nvim_lsp.jsonls.setup(final_opts)
 		end,
 
-		sumneko_lua = function()
-			local _opts = require("completion.servers.sumneko_lua")
+		lua_ls = function()
+			local _opts = require("completion.servers.lua_ls")
 			local final_opts = vim.tbl_deep_extend("keep", _opts, opts)
-			nvim_lsp.sumneko_lua.setup(final_opts)
+			nvim_lsp.lua_ls.setup(final_opts)
 		end,
 	})
